fastapi/inst/requirements.txt
uvicorn
fastapi
sqlalchemy
pymysql
passlib
python-multipart
aiofiles
pyjwt